Combine send and recv test
