+++ test/test_using_simulaqron.sh
@@ -22,20 +22,19 @@ simulaqron set network-config-file $network_file
 # Change directory as well so this script work from wherever
 cd $this_folder_path
 
+# Build the tests
+echo -e "\e[1;32m[$(date +%H:%M:%S)] Build Tests\e[0m"
+make all
+
 # Start the Virtual and CQC Nodes
 echo -e "\e[1;32m[$(date +%H:%M:%S)] Start CQC Nodes\e[0m"
 simulaqron start --nodes="Alice,Bob" --keep
 sleep 5
 
-# Build the tests
-echo -e "\e[1;32m[$(date +%H:%M:%S)] Build Tests\e[0m"
-make all
-
 # Start the tests
 echo -e "\e[1;32m[$(date +%H:%M:%S)] Run Tests\e[0m"
 bin/qubit localhost 8803
-bin/send localhost 8803 localhost 8804
-bin/recv localhost 8804
+bin/send_recv localhost 8803 localhost 8804
 bin/gates localhost 8803
 
 echo -e "\e[1;32m[$(date +%H:%M:%S)] Testing Complete\e[0m"
